+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -7,11 +7,12 @@ Release History
 +++++
 * BREAKING CHANGE: Remove 'acr build-task' command group.
 * BREAKING CHANGE: Remove --tag/--manifest from 'acr repository delete' command.
+* Add '--target' parameter for 'az acr build', 'az acr task create' and 'az acr task update' commands.
 
 2.1.13
 ++++++
 * Add support for VNet/IP rules.
-* Add --platform parameter for `acr build`, `acr task create` and `acr task update` commands
+* Add '--platform' parameter for 'acr build', 'acr task create' and 'acr task update' commands.
 
 2.1.12
 ++++++